+++ src/imports/completions/conditionsCompletionProvider.ts
@@ -5,11 +5,11 @@ import { addMechanicCompletions } from '../utils/completionhelper';
 
 export function conditionCompletionProvider(){
     const conditionCompletionProvider = vscode.languages.registerCompletionItemProvider(
-        'yaml',
+        'mythicscript',
         {
             async provideCompletionItems(document: vscode.TextDocument, position: vscode.Position, _token: vscode.CancellationToken, context: vscode.CompletionContext) {
     
-                const keys = yamlutils.getParentKeys(document, position.line);
+                const keys = yamlutils.getParentKeys(document, position);
                 if (!keyAliases["Conditions"].includes(keys[0])) {
                     return undefined;
                 }