+++ src/imports/completions/filecompletions/triggerfileCompletionProvider.ts
@@ -3,14 +3,13 @@ import * as yamlutils from '../../utils/yamlutils';
 import { keyAliases, Mechanic, ObjectInfo, ObjectType } from '../../../objectInfos';
 import { checkShouldComplete } from '../../utils/completionhelper';
 
-
 export function triggerfileCompletionProvider(){
     const triggerfileCompletionProvider = vscode.languages.registerCompletionItemProvider(
-        'yaml',
+        'mythicscript',
         {
             async provideCompletionItems(document: vscode.TextDocument, position: vscode.Position, _token: vscode.CancellationToken, context: vscode.CompletionContext) {
         
-                const keys = yamlutils.getParentKeys(document, position.line);
+                const keys = yamlutils.getParentKeys(document, position);
                 if (!keyAliases["Skills"].includes(keys[0])) {
                     return undefined;
                 }