+++ src/imports/completions/inlinemetaskillCompletionProvider.ts
@@ -6,11 +6,11 @@ import { keyAliases } from '../../objectInfos';
 export function inlineMetaskillCompletionProvider(){
 
     const inlineMetaskillCompletionProvider = vscode.languages.registerCompletionItemProvider(
-        'yaml',
+        'mythicscript',
         {
             async provideCompletionItems(document: vscode.TextDocument, position: vscode.Position) {
     
-                const keys = yamlutils.getParentKeys(document, position.line);
+                const keys = yamlutils.getParentKeys(document, position);
                 if (!keyAliases["Skills"].includes(keys[0])) {
                     return undefined;
                 }