+++ src/imports/completions/mechaniclineCompletionProvider.ts
@@ -8,7 +8,7 @@ import { keyAliases } from '../../objectInfos';
 export function mechaniclineCompletionProvider(){
 
     const mechaniclineCompletionProvider = vscode.languages.registerCompletionItemProvider(
-        'yaml',
+        'mythicscript',
         {
             async provideCompletionItems(document: vscode.TextDocument, position: vscode.Position) {
     
@@ -17,7 +17,7 @@ export function mechaniclineCompletionProvider(){
                     return undefined;
                 }
     
-                const keys = yamlutils.getParentKeys(document, position.line);
+                const keys = yamlutils.getParentKeys(document, position);
                 if (!keyAliases["Skills"].includes(keys[0])) {
                     return undefined;
                 }